Prepending every reference to a variable or function declared in harmonyOscillatorsGlobalNamespace with "harmonyOscillatorsGlobalNamespace.".
NOTE: Obviously, this makes the code much less readable, but it is a clean refactor (i.e., nothing is broken). Ideally, if function x() and y() were both in the namespace object, they would call each other by this.x() and this.y(), but the current code has a lot of function nesting (closures) and callback functions, which can make the value of "this" less obvious. The next few commits can start switching things to "this." instead of "harmonyOscillatorsGlobalNamespace." one-by-one to make sure things don't break and thereby determine which functions and variables actually need to be accessed directly from the harmonyOscillatorsGlobalNamespace global variable.